+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Utils.hs
@@ -7,6 +7,7 @@ module Test.Cardano.Ledger.Core.Utils (
   testGlobals,
   mkDummySafeHash,
   txInAt,
+  mkActiveStake,
 ) where
 
 import Cardano.Ledger.BaseTypes (
@@ -15,13 +16,21 @@ import Cardano.Ledger.BaseTypes (
   Network (..),
   knownNonZeroBounded,
   mkActiveSlotCoeff,
+  nonZero,
  )
+import Cardano.Ledger.Coin (Coin)
+import Cardano.Ledger.Compactible (toCompactPartial)
 import Cardano.Ledger.Core
+import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.Hashes (unsafeMakeSafeHash)
+import Cardano.Ledger.State (ActiveStake (..), StakeWithDelegation (..))
 import Cardano.Ledger.TxIn (TxIn, mkTxInPartial)
 import Cardano.Slotting.EpochInfo (fixedEpochInfo)
 import Cardano.Slotting.Time (SystemStart (..), mkSlotLength)
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
+import qualified Data.VMap as VMap
 import Test.Cardano.Ledger.Binary.Random (mkDummyHash)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Rational (unsafeBoundRational)
@@ -49,3 +58,16 @@ txInAt :: (HasCallStack, EraTx era) => Int -> Tx l era -> TxIn
 txInAt index tx =
   let txId = txIdTx tx
    in mkTxInPartial txId (toInteger index)
+
+-- | Construct an @ActiveStake@ from separate maps of stake and delegations.
+-- Only credentials present in both maps with non-zero stake are included.
+mkActiveStake ::
+  Map (Credential Staking) Coin ->
+  Map (Credential Staking) (KeyHash StakePool) ->
+  ActiveStake
+mkActiveStake stakeMap delegsMap =
+  ActiveStake
+    $ VMap.fromMap
+    $ Map.mapMaybe
+      (\(c, d) -> flip StakeWithDelegation d <$> nonZero (toCompactPartial c))
+    $ Map.intersectionWith (,) stakeMap delegsMap