+++ libs/cardano-ledger-test/bench/Bench/Cardano/Ledger/StakeDistr.hs
@@ -234,7 +234,7 @@ tickfRuleBench =
                     , bgroup
                         "calculatePoolDistr subparts"
                         [ bench "poolStake" $
-                            whnf (calculatePoolStake (const True) (ssDelegations (getSnap nes))) (ssActiveStake (getSnap nes))
+                            whnf (calculatePoolStake (const True)) (ssActiveStake (getSnap nes))
                         , bench "old calculatePoolDistr" $
                             whnf (oldCalculatePoolDistr (const True)) (getSnap nes)
                         ]