+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.21.0.0
 
+* Update `resolveConwayInstantStake` to return `ActiveStake`
 * Remove `asBoundedIntegralHKD`
 * Change lens type of `hkdCommitteeMinSizeL` and `ppCommitteeMinSize` from `Natural` to `Word16`
 * Add `HeaderProtVerTooHigh` predicate failure.